refactor: rename DisplayHelper.kt and update package structure to core
