+++ library/src/main/java/com/github/only52607/compose/core/DisplayHelper.kt
@@ -1,4 +1,4 @@
-package com.github.only52607.compose.window
+package com.github.only52607.compose.core
 
 
 import android.content.Context
